Major refactoring in a lot of places. Completed the shooting code for Muka.

+++ core/src/rami/project/grey/core/entity/ILiveable.java
@@ -4,11 +4,11 @@ package rami.project.grey.core.entity;
  * Describes an entity that has health and can attack. The units are in HP.
  * */
 public interface ILiveable extends IEntity {
-    short getTotalHealth();
-    short getCurrentHealth();
+    int getTotalHealth();
+    int getCurrentHealth();
 
     /**
      * @return how much HP should be deducted by an attack
      * */
-    short getDamageDealt();
+    int getDamageDealt();
 }
